cloud/cloud_function/requirements.txt
functions-framework
google-cloud-storage
google-cloud-pubsub
google-cloud-secret-manager
google-cloud-aiplatform
pillow

